refactor: a bit of refactoring
